requirements.txt
Flask 
requests
python-dotenv
httpx==0.17.0